refactor: Stations / Lines 일급 컬렉션 제거
